pyproject.toml
[project]
name = "demo-mcp-github"
version = "1.0.0"
description = "Demo MCP server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.72.1",
    "arxiv>=2.3.1",
    "fastapi>=0.121.2",
    "fastmcp>=2.13.0.2",
    "mcp>=1.21.0",
    "python-dotenv>=1.2.1",
    "uv>=0.9.9",
]
